refactor(lex-analyzer): organizing directories
