+++ README.md
@@ -1,2 +1,15 @@
 # tiger-compiler
+
 Compiler build in C for the Tiger language
+
+Linux users should install the follow dependencies:
+`apt install flex`
+
+MacOs users should install the follow dependencies:
+`brew install flex`
+
+To compile the project:
+`make`
+
+To execute:
+`./tiger examples/{example_name}`
