+++ examples/example1
@@ -1,0 +1,5 @@
+let
+    // comentario teste
+    if (1 < 2){
+        helloWorld := "Hello, World!"
+    }
