+++ include/utilities.h
@@ -1,0 +1,7 @@
+#include <assert.h>
+
+typedef char *string;
+typedef char bool;
+
+void *check_malloc(int);
+string String(char *);
