+++ src/utilities.c
@@ -1,4 +1,4 @@
-#include "utilities.h"
+#include "../include/utilities.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -16,12 +16,4 @@ string String(char *s){
     string p = check_malloc(strlen(s)+1);
     strcpy(p,s);
     return p;
-}
-
-U_boolList U_BoolList(bool head, U_boolList tail)
-{ 
-    U_boolList list = check_malloc(sizeof(*list));
-    list->head = head;
-    list->tail = tail;
-    return list;
 }
