--- utilities.h
@@ -1,13 +1,0 @@
-#include <assert.h>
-
-typedef char *string;
-typedef char bool;
-
-void *check_malloc(int);
-string String(char *);
-
-typedef struct U_boolList_ *U_boolList;
-struct U_boolList_ {bool head; U_boolList tail;};
-U_boolList U_BoolList(bool head, U_boolList tail);
-
-
